Cambios de rosa
Productoservlet vista

+++ SysSeguridad.EntidadesDeNegocio/src/sysseguridad/entidadesdenegocio/Producto.java
@@ -4,7 +4,6 @@
  */
 package sysseguridad.entidadesdenegocio;
 
-import java.math.BigDecimal;
 
 
 public class Producto {
@@ -14,7 +13,7 @@ public class Producto {
     private int IdProveedor;
     private String Nombre;
     private String Descripcion;
-    private BigDecimal Precio;
+    private Float Precio;
     private int Existencia;
     private int top_aux;
     private Categoria Categoria;
@@ -23,7 +22,7 @@ public class Producto {
     public Producto() {
     }
 
-    public Producto(int IdProducto, int IdCategoria, int IdProveedor, String Nombre, String Descripcion, BigDecimal Precio, int Existencia) {
+    public Producto(int IdProducto, int IdCategoria, int IdProveedor, String Nombre, String Descripcion, Float Precio, int Existencia) {
         this.IdProducto = IdProducto;
         this.IdCategoria = IdCategoria;
         this.IdProveedor = IdProveedor;
@@ -73,11 +72,11 @@ public class Producto {
         this.Descripcion = Descripcion;
     }
 
-    public BigDecimal getPrecio() {
+    public Float getPrecio() {
         return Precio;
     }
 
-    public void setPrecio(BigDecimal precio) {
+    public void setPrecio(Float precio) {
         this.Precio = precio;
     }
